+++ src/main/java/lotto/hadler/ExceptionHandler.java
@@ -2,9 +2,26 @@ package lotto.hadler;
 
 import java.util.function.Supplier;
 
-public interface ExceptionHandler {
+public class ExceptionHandler {
 
-    <T> T trySupplierUntilSuccess(Supplier<T> supplier);
+    public <T> T trySupplierUntilSuccess(Supplier<T> supplier) {
+        while (true) {
+            try {
+                return supplier.get();
+            } catch (IllegalArgumentException e) {
+                System.out.println(e.getMessage());
+            }
+        }
+    }
 
-    void tryRunnableUntilSuccess(Runnable runnable);
+    public void tryRunnableUntilSuccess(Runnable runnable) {
+        while (true) {
+            try {
+                runnable.run();
+                break;
+            } catch (IllegalArgumentException e) {
+                System.out.println(e.getMessage());
+            }
+        }
+    }
 }
